User commit

+++ src/main/java/com/example/demo/repository/ValidationRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.entity.ValidationEntity;
 
 @Repository
-public interface ValidationRepository extends JpaRepository<ValidationEntity,Integer>{
+public interface ValidationRepository extends JpaRepository<ValidationEntity,Long>{
        
 }
